Ensure that Antora can build the generated site
Recent versions of `actions/upload-artifact@v4` don't include hidden
files by default anymore. This causes an error in the GitHub workflow
since Antora refuses to render a site which isn't stored in a Git
repository.

This commit ensures that hidden files are included in the artifact
that's used to share the generated Antora sources between the workflow
jobs.